fix(HumanVsEngine): bug in undo and switch sides
Undo-ing moves was not forwared to the chess engine. That has been fixed. It seems event handlers (in our case onmessage handler) cannot access react state - they only access the value that has been set during initialization. This caused a problem as onmessage handler was using playerColor state value. To fix this playerColorRef is created and is used.